Add CheckInPresenterTests and refactored CheckInPresenter to include value checks.
The value checks ensure that the user is not going to enter a character or long number that will break the app.
The tests were then added; they test all the possibilities.

As of this commit, the app is complete. I will commit the release version of the app next.